+++ gh-pages-clean.js
@@ -1,0 +1,4 @@
+
+import * as ghpages from 'gh-pages';
+
+ghpages.clean();